pyproject.toml
[project]
name = "agenttriagem"
description = "Sistema Híbrido de Agentes de IA para Triagem Médica - Detecção de Pneumonia"
version = "0.1.0"
packages = [{include = "src"}]
license = "MIT"
authors = [
    {name = "vinicius.miranda",email = "vncs2012@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "google-adk (>=1.22.1,<2.0.0)",
    "litellm (>=1.81.0,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "fastapi[standard] (>=0.115.0,<0.124.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "mcp (>=1.25.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"








[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]
